third_party/WebKit/LayoutTests/inspector/console/console-dir-expected.txt

CONSOLE MESSAGE: line 9: test1,test2
CONSOLE MESSAGE: line 10: [object NodeList]
CONSOLE MESSAGE: line 11: [object XPathResult]
CONSOLE MESSAGE: line 20: [object Object]
Tests that console logging dumps proper messages.



Array[2]
    0: "test1"
    1: "test2"
    length: 2
    __proto__: Array[0] console-dir.html:9

NodeList[1]
    0: html
    length: 1
    __proto__: NodeList console-dir.html:10

XPathResult
    booleanValue: [Exception: TypeError]
    invalidIteratorState: false
    numberValue: [Exception: TypeError]
    resultType: 4
    singleNodeValue: [Exception: TypeError]
    snapshotLength: [Exception: TypeError]
    stringValue: [Exception: TypeError]
    __proto__: XPathResult console-dir.html:11

Object
    "": ""
    "  ": "  "
    " a b ": " a b "
    $foo5_: 0
    "a↵↵b↵c": "a↵↵b↵c"
    c d: "c d"
    __proto__: Object console-dir.html:20
